importing a complete copy of the frontend folder comments from https://github.com/BarnesFoundation/WebsiteRedesign
This folder contains components and styles from the barnes style guide
toolkit. It was copied over from this commit

https://github.com/BarnesFoundation/WebsiteRedesign/commit/cc262a553eb8b
313af5acde332dd343e2d875aed

Note that these html files are twig template files. These will need to
be converted to jsx. We might also cull out components that we don’t
need.

Also note that this doesn’t build yet because of an error:

`Module not found: You attempted to import
../fonts/CalibreWeb-Regular.woff2 which falls outside of the project
src/ directory. Relative imports outside of src/ are not supported. You
can either move it inside src/, or add a symlink to it from project's
node_modules/.` This will be fixed in the next commit.