Use latest timestamped indices instead of hardcoded collection_{a,b}
Instead of manually specifying collection_{a,b} sides in
ELASTICSEARCH_INDEX, we just look at the latest collection_<timestamp>
index that has at least one record with a tag field.

The tag field is populated in the last step of the update scripts
in barnes-tms-extract, so its a good indicator that this index is
complete.